# trabaio/backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API do Sistema de Petshop",
    "version": "1.0.0",
    "description": "API para gerenciamento de petshop com autenticação JWT"
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Servidor de Desenvolvimento"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "AdminAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/products": {
      "get": {
        "summary": "Listar todos os produtos",
        "tags": ["Produtos"],
        "security": [
          {"BearerAuth": []}
        ],
        "responses": {
          "200": {
            "description": "Lista de produtos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Criar novo produto",
        "tags": ["Produtos"],
        "security": [
          {"BearerAuth": []}
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductCreation"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Produto criado com sucesso"
          },
          "500": {
            "description": "Erro ao cadastrar o produto"
          }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "summary": "Buscar produto por ID",
        "tags": ["Produtos"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {"type": "integer"},
            "description": "ID do produto"
          }
        ],
        "responses": {
          "200": {
            "description": "Dados do produto",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "404": {"description": "Produto não encontrado"},
          "500": {"description": "Erro ao buscar produto"}
        }
      },
      "put": {
        "summary": "Atualizar produto por ID (completo)",
        "tags": ["Produtos"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {"type": "integer"},
            "description": "ID do produto"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductCreation"
              }
            }
          }
        },
        "responses": {
          "200": {"description": "Produto atualizado com sucesso"},
          "404": {"description": "Produto não encontrado"},
          "500": {"description": "Erro ao atualizar o produto"}
        }
      },
      "patch": {
        "summary": "Atualizar produto por ID (parcial)",
        "tags": ["Produtos"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {"type": "integer"},
            "description": "ID do produto"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductCreationPartial"
              }
            }
          }
        },
        "responses": {
          "200": {"description": "Produto atualizado parcialmente com sucesso"},
          "404": {"description": "Produto não encontrado"},
          "500": {"description": "Erro ao atualizar o produto parcialmente"},
          "delete": {
            "summary": "Deletar produto por ID",
            "tags": ["Produtos"],
            "security": [
              {"BearerAuth": []}
            ],
            "parameters": [
              {
                "in": "path",
                "name": "id",
                "required": true,
                "schema": {"type": "integer"},
                "description": "ID do produto"
              }
            ],
            "responses": {
              "200": {"description": "Produto deletado com sucesso"},
              "404": {"description": "Produto não encontrado"},
              "500": {"description": "Erro ao deletar o produto"}
            }
          }
        }
      }
    },
    "/services": {
      "get": {
        "summary": "Listar todos os serviços",
        "tags": ["Serviços"],
        "security": [
          {"BearerAuth": []}
        ],
        "responses": {
          "200": {
            "description": "Lista de serviços",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Service"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Criar novo serviço",
        "tags": ["Serviços"],
        "security": [
          {"AdminAuth": []}
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceCreation"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Serviço criado com sucesso"
          },
          "500": {
            "description": "Erro ao criar o serviço"
          }
        }
      }
    },
    "/services/{id}": {
      "get": {
        "summary": "Buscar serviço por ID",
        "tags": ["Serviços"],
        "security": [
          {"BearerAuth": []}
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {"type": "integer"},
            "description": "ID do serviço"
          }
        ],
        "responses": {
          "200": {
            "description": "Dados do serviço",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Service"
                }
              }
            }
          },
          "404": {"description": "Serviço não encontrado"},
          "500": {"description": "Erro ao buscar o serviço"}
        }
      },
      "put": {
        "summary": "Atualizar serviço por ID (completo)",
        "tags": ["Serviços"],
        "security": [
          {"AdminAuth": []}
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {"type": "integer"},
            "description": "ID do serviço"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceCreation"
              }
            }
          }
        },
        "responses": {
          "200": {"description": "Serviço atualizado com sucesso"},
          "404": {"description": "Serviço não encontrado"},
          "500": {"description": "Erro ao atualizar o serviço"}
        }
      },
      "patch": {
        "summary": "Atualizar serviço por ID (parcial)",
        "tags": ["Serviços"],
        "security": [
          {"AdminAuth": []}
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {"type": "integer"},
            "description": "ID do serviço"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceCreationPartial"
              }
            }
          }
        },
        "responses": {
          "200": {"description": "Serviço atualizado parcialmente com sucesso"},
          "404": {"description": "Serviço não encontrado"},
          "500": {"description": "Erro ao atualizar o serviço parcialmente"},
          "delete": {
            "summary": "Deletar serviço por ID",
            "tags": ["Serviços"],
            "security": [
              {"AdminAuth": []}
            ],
            "parameters": [
              {
                "in": "path",
                "name": "id",
                "required": true,
                "schema": {"type": "integer"},
                "description": "ID do serviço"
              }
            ],
            "responses": {
              "200": {"description": "Serviço deletado com sucesso"},
              "404": {"description": "Serviço não encontrado"},
              "500": {"description": "Erro ao deletar o serviço"}
            }
          }
        }
      }
    },
...
  }
}